Add key figures and tables from country plans
