+++ src/App/routes/CountryRoutes.tsx
@@ -437,19 +437,6 @@ const countryPreparedness = customWrapRoute({
     },
 });
 
-const countryPlan = customWrapRoute({
-    parent: countriesLayout,
-    path: 'plan',
-    component: {
-        render: () => import('#views/CountryPlan'),
-        props: {},
-    },
-    context: {
-        title: 'Country Plan',
-        visibility: 'anything',
-    },
-});
-
 const countryAdditionalInfo = customWrapRoute({
     parent: countriesLayout,
     path: 'additional-info',
@@ -497,5 +484,4 @@ export default {
     countryThreeWIndex,
     countryRiskWatch,
     countryPreparedness,
-    countryPlan,
 };
