+++ src/views/CountryNsOverviewStrategicPriorities/StrategicPrioritiesTable/i18n.json
