+++ src/views/CountryNsOverviewStrategicPriorities/styles.module.css
@@ -1,9 +1,11 @@
 .country-plan {
     padding: var(--go-ui-spacing-2xl) 0;
 
-    .actions {
+    .country-plan-download-link {
+        display: flex;
         flex-shrink: unset;
         flex-wrap: wrap;
+        gap: var(--go-ui-spacing-lg);
 
         .icon {
             font-size: var(--go-ui-height-icon-multiplier);
@@ -12,15 +14,15 @@
 
     .content {
         display: flex;
-        flex-direction: column;
-        gap: var(--go-ui-spacing-2xl);
-        padding: var(--go-ui-spacing-xl) 0;
+        align-items: center;
+        flex-wrap: wrap;
+        gap: var(--go-ui-spacing-3xl);
+        justify-content: space-evenly;
+        padding: var(--go-ui-spacing-md) 0;
 
         .key-figures {
             display: flex;
-            flex-wrap: wrap;
             gap: var(--go-ui-spacing-md);
-            justify-content: center;
 
             .key-figure {
                 flex-basis: 16rem;
@@ -29,16 +31,10 @@
             }
         }
 
-        .tables-section {
-            display: flex;
-            gap: var(--go-ui-spacing-xl);
-            flex-wrap: wrap;
-
-            .strategic-priority-table,
-            .membership-coordination-table {
-                flex-grow: 1;
-                overflow-x: auto;
-            }
+        .strategic-priority-table {
+            flex-grow: 1;
+            width: 50%;
+            overflow-x: auto;
         }
     }
 }
