--- src/views/CountryPlan/MembershipCoordinationTable/i18n.json
@@ -1,7 +1,0 @@
-{
-    "namespace": "countryPlan",
-    "strings": {
-        "countryPlanMembershipCoordinationTableTitle": "Membership Coordination",
-        "countryPlanNameOfPNS": "Name of Partner National Society"
-    }
-}
